Add pivot table docs, minor stylistic code changes
The pivot table function is cool :-) I added a few samples to the docs.
I also did a few minor changes (as I was reading the code) - there
was absolutely nothing wrong with it - I did just a few minor stylistic
tweaks and adjusted it to a small notebook screen!