More systematic naming for ordering/sorting functions
Previously, we had `orderCols` and `orderRows`, but since all the new sorting functions
use "sort" in the name (which I think is a good idea) and there is a wide range of
`sortRows`, `sortRowsBy` and `sortRowsWith` functions (all well named), this commit
makes `orderRows` obsolete and replaces it with `sortRowsByIndex` (and similarly for
columns, to keep the symmetry).

I left the obsolete functions there to simplify migration, but marked them obsolete.
We can remove them in some future version.