Refactor docs, add formatters for series, frame and charts
This checkin does not change anything in the implementation (it only
removes obsolete Series.mean etc. functions). Aside from that, it reorganizes
and improves documentation in a number of ways.

The most fun part is that F# fornmatting now lets us automatically embed
series, frames and charts, so this is used in a couple of places :-)