Cleanup binomial heap and use it only when appending larger number of series.
I did some cleanup in the binomial heap code, but have not changed any logic
(and I added one simple FsCheck test to check that it works OK).

I did some performance evaluation and the binomial heap implementation is much
faster for large number of series - usually around 150 (regardless of their length),
but for smaller number of series, it tends to be 2x slower. So I left both
implementations in the codebase and added a function that chooses the right one
at runtime. This is not entirely nice, but it seems to give us the best perf!