Do more renaming ("series" and "column") (c.f. #189)
I found a few more occurrences of "series" and replaced them with
"column" to finish the idea in PR #190.

There was another minor inconsistency where we sometimes used
`cols` and sometimes `columns` in the F# API, so this changes
everything (in the F# module only) to use `cols`. I would be equally
happy with `columns` though (I just picked the more common one).