Speedup df.AppendN (and rename appending functions on frame to merging)
There was a failing statistics test, but this was due to floating point
precision issue, so I updated the test.

The last failing test is the one comparing our skewness with Math.NET
but that is a Math.NET issue (#199)

I have not yet renamed Combine/CombineN in vector construction. I would
like to get rid of Combine and have just a single Merge, but I want to
see if this has any performance implications for two-element mergings
(which is probably the common case).