Merge new deque implementation and some speedup
The new Deque is slightly slower (say 10%), but there was enough room for
other optimizations elsewhere, so I made a couple of minor tweaks in Stats
module and so the overall result is now faster. Not sure what is different
in our F# queue - it is likely just my use of Debug build.