Documentation and API improvements, fix indexWith function
This adds remaining API documentation for the `Series` module and
rearranges functions in the `Frame` module. I have not done any actual
changes - just reordered functions and added comments. The goal is to have
nice generated documentation for the functions in these two modules. The
checkin also renamed "categories" in the two files, so that they mostly
match and make sense. Documentation for `Stats` module looks nice too.

I also renamed `Frame.reduce` and `Series.reduce` rename to `reduceValues`
to keep this in sync with the naming elsewhere (without `Values`, I'd expect
it to call the specified function with the key as wall as the value).
I added `fold` operation to series for symmetry too.

This fixes one bug with the `indexWith` operation too. Previously, the operation
just used the original vector, which worked well only when the new sequence of
keys was of the same length as the vector. This is now fixed (the operation
either pads the vector or drops a part of it).