Fix chunking bug and add Travis build
The Series.chunkSize function had a bug when the chunk at the boundary was empty.
This is now fixed and I added a test to cover this.

The rest of the commit adds integration with the Travis CI server (which works
nicely with GitHub and is used by other F# open-source projects). It builds and
runs the project on Mono on a Mac, so I had to do some fixes to make things work
on Mono. Building and running tests for R interop is still TODO, but the rest works
nicely.

I also improved README file and added "Build passing/failing" icon there.