Update documentation for the Frame module
This finishes the cleanup of the `Frame` module so that we can
automatically generate a nice documentation from the comments
(i.e. http://bluemountaincapital.github.io/Deedle/reference/deedle-framemodule.html).
This also updates the F# Formatting references and performs other
documentation tweaks elsewhere.

The only API change is that I renamed `getNumericColumns` to
`getNumericCols` (to be consistent with the rest of the module,
where we always use the "cols" shortcut).