Documentation and change boundary behavior in expanding means
This includes some final tweaks in the documentation (cleanup the home page a little
bit, make it more similar to R provider and F# Data, update F# Formatting) and it
adds documentation for the `Stats` module.

While working on the docs, I noticed that `Stats.expandingMean` returns zero when
the series starts with N/A. For example:

    series [ 0 => nan; 1 => 1.0 ] |> Stats.expandingMean

... was returning `series [ 0 => 0.0; 1 => 1.0 ]`. This looks wrong to me (calculating
the mean of a window containing no values, should be N/A), so I chnaged the behavior to
return `series [ 0 => nan; 1 => 1.0 ]`.

Another slightly surprising fact I noticed is that `movingMin` (or max) and `movingMean`
(or other statistical functions) have different boundary behavior. I'm not sure if this
is intentional or not - moving min/max find value even for
"incomplete" windows at the beginning of the series (using a windows of size 1 to (n-1)).
On the other hand, stats functions just return N/A for the first n-1 keys. (This is
equivalent to the difference between `Series.windowSize (1, Boundary.Skip)` (which
is how stats functions behave) and `Series.windowSize (1, Boundary.AtBeginning)` (which
is how min/max behave). I left this as it is, but explained it in the documentation.