Typed access to Rows via an automatically implemented interface
Initial implementation that makes it possible to use frame rows in a more
type safe way than using just `df.Rows` (without waiting for type provider
features in the next version of F#).

This lets the user define an interface (by hand) and then fits the columns
of the frame to the members of the interface (see example in tests).