Support adding/replacing columns with same-indexed series in BigDeedle
In general, we do not (yet?) support joins on BigDeedle frames and series.
But if the newly added (joined) frame or series has the same index, then
this should work. But - the index might not even support "GetRange" (which
is normally called before testing equality), so we first check if they are
the same object (which is a useful optimization anyway).

In the future, it would make sense to have some mechanism for testing
which operations are supported (so that we do not try calling
"restrict" if it is not supported). But for now, equality test should do.