Fix R provider test issue after Paket conversion
The problem was that one of the test projects referenced RProvider even though
it did not use it (which failed as R provider does not work on Travis at the
moment). This fixes that, reruns "paket install", adds project order requirement
(wihch was missing) to Deedle.Core.sln and also adds workaround to `rinterop.fsx`
(which will be fixed once we update R provider).