Handle F# Data dependency (for CSV parsing) via Github
This removes the dependency on FSharp.Data package. We can now reference
the files we need directly from F# Data Github using a new nifty Paket feature.

The benefit is that we have a smaller stand-alone package that can be
combined with any version of F# Data and also is less prone to binding redirect
errors (no MethodMissingException when using wrong FSharp.Core.dll).