Dropping sparse rows/columns should preserve frame structure (fix #277)
This is implemented differently for rows and columns. For rows, we should be
able to make this work for virtual frames - the implementation is similar to
`frame.Rows` in that it creates a (virtual) vector containing `true`/`false`
whether the row is empty - and then we perform (scan) using `Search`.

For columns, we just do a simple array processing (we have to walk over the
whole column anyway, so this would not work for virtual series).