Enable materializing delayed series into a virtual series
This was surprisingly easy change. Previously, the `loader` for `DelayedSeries` was
a function that returned key value  pairs and the delayed series then turned this
into vector and index. This change adds an overload where `loader` returns index and
vector directly - and so we can create loader that returns virtual frame & series
(and we also need to make sure that Materialize uses the new index builder).

I renamed the original `DelayedSeries.Create` to `DelayedSeries.FromValueLoader`
and added `DelayedSeries.FromIndexVectorLoader` (the name is a bit silly - better
suggestions are welcome!)