Enforce unique keys when series is accessed (fix #294)
In a recent change, we started initializing the lookup table for in-memory
indices lazily (to make certain trivial operations faster). But this means
we did not check for unique keys in the index on creation.
This adds the check when accessing all keys, key count etc. - we are still
somewhat lazy, but check keys when the series is accessed in any reasonable
way (most importantly - including printing - so you'll see the error in FSI).