Printing accesses only first and last partitions
This changes how printing of a series and frame works (and also operations
like take & takeLast), so that they do not need the total number of elements
in the series. The trick is to change `GetAddressRange` so that it does not
take just `lo, hi`, but can also take `Start(count)` and `End(count)`.
Aside from that, a few changes in how `Format` is implemented is needed
(e.g. we cannot check that `KeyCount < startPrint + endPrint`, but instead,
we need to skip `startPrint + endPrint` items and then see if there is more).