Abstracts how location in a vector is calculated
When calling `series.Select`, the function has access to the offset within the series. This is
tricky for BigDeedle where projections are done lazily - when we need to evaluate value for a
given key, we need to calculate offset of the key in the series.

This commit introduces `IVectorLocation` abstraction that lets us do most operations without
actually calculating the offset (the offset is only calculated when the user actually uses
`Select` overload taking index).