Don't iterate sequence twice in Frame.ofRowsOrdinal
`Frame.ofRowsOrdinal` iterates the sequence TWICE. This is a big issue
when the implementation of the sequence involves some resources at may
only be able to be iterated once, for example, an IDataReader instance
from a database connection.

While it can be worked around by caching the sequence into an array
before feeding it to Frame.ofRowsOrdinal. But in general, the common
practice of working with seq and IEnumerable is that the function that
takes the sequence (the callee) should take care of caching if it is to
be iterates multiple times.