Don't iterate sequence twice in Frame.ofRowsOrdinal
This addresses issue #356

`Frame.ofRowsOrdinal` iterates the sequence TWICE. This is a big issue
when the implementation of the sequence involves some resources at may
only be able to be iterated once, for example, an IDataReader instance
from a database connection.

The fix is to only iterate it once in `Frame.ofRowsOrdinal`.